Removed docker hub image & update info
Since we moved to docker-compose using docker hub for the image isn't useful as far as I know, in the future we can do this again or add more details about docker-compose but until then I'm leaving it to the user to learn how to use docker-compose. Mostly since Its new to me.

The docker image I put on docker hub is no longer available.